fast_api/requirements.txt
h11==0.14.0
annotated-types==0.7.0
pydantic_core==2.23.4
greenlet==3.1.1
sniffio==1.3.1
starlette==0.38.6
uvicorn==0.31.0
anyio==4.6.0
fastapi==0.115.0
redis==5.1.0
pydantic==2.9.2
typing_extensions==4.12.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
asyncpg==0.29.0
async-timeout==4.0.3
PyJWT==2.9.0
aioredis==2.0.1
